// FO/gestpat-frontend/src/app/models/public-pat.model.ts
export class PublicPat {
  id?: string;
  registrationNumber?: string;
  type?: string;
  typeAr?: string;
  landReferences?: string;
  landReferencesAr?: string;
  registrationReference?: string;
  registrationDate?: string;
  area?: number;
  location?: string;
  locationAr?: string;
  latitude?: number;
  longitude?: number;
  acquisitionSource?: string;
  acquisitionSourceAr?: string;
  purchasePrice?: number;
  legalBasis?: string;
  legalBasisAr?: string;
  zoningDesignation?: string;
  zoningDesignationAr?: string;
  currentUse?: string;
  currentUseAr?: string;
  marketValue?: number;
  isPrivatelyUsedByThirdParty?: boolean;
  privateUseDetails?: string;
  privateUseDetailsAr?: string;
  occupationPermitDate?: string;
  authorizedPerson?: string;
  authorizedPersonAr?: string;
  temporaryOccupationDuration?: number;
  occupationFeeDuration?: number;
  occupationFeeAmount?: number;
  increaseRatePercent?: number;
  paymentDeadlinesAndMethods?: string;
  paymentDeadlinesAndMethodsAr?: string;
  removalFromPublicDomainReference?: string;
  removalFromPublicDomainDate?: string;
  notes?: string;
  notesAr?: string;
}